requirements.txt
# RTX 3050 4GB Grain Quality Assessment Requirements
# Install with: pip install -r requirements.txt

# Core ML/DL packages
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Data processing
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.2.0

# Image processing
Pillow>=9.0.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Excel file support
openpyxl>=3.0.0

# Progress bars and utilities
tqdm>=4.64.0

# Optional: For enhanced performance monitoring
# psutil>=5.9.0
# GPUtil>=1.4.0